feat: P0+P1+P2 全量补齐 — 仓颉 iOS 移植客户端对齐 PlotPilot v4.6.0
P0（9项核心砍功能）：
- 4个章节生成SSE端点 + WorkbenchStore.analyzeScene/retrieveContext
- DAGRunStore.latestResult + DAG SSE性能优化层
- Tools.tensionSlingshot + 6项模型解码修复

P1（6任务重要缺失）：
- 38端点 + 18 query + 4路径修正
- 15新struct + 8模块字段补齐
- AutopilotWorkspaceStore 4Tab + StatsStore 3缓存
- 7项视图组件 + ChapterContentPanel重命名

P2（25项体验降级+遗留清零）：
- 视图层8项（charts/global/onboarding/brand/LogPanel/StylePreset/DialogueCorpus）
- 交互3项（Home批量/ChapterList视图切换/Autopilot懒加载）
- 模型层3项（GenerationPrefsDTO/CastCoverage/EvolutionGateReport.violations/BookStats.avgChapterWords）
- 状态层3项（generationPollMs/guardrail退避/pauseWhenHidden）
- P1遗留5项清零（toChapterMicroBeatPayloads/microBeats接线/triples接线/streamPhase去重/端点消费者）

修复编译隐患：
- APIEndpoint.swift Memory enum 重复case
- LogPanelView LogEntry重命名为LogPanelEntry
- OnboardingWizardView 删除重复EmptyBody定义

三批全过v2门禁验收（死API按U2/U3/U4决策跳过，不发明功能）。

+++ Cangjie/Models/StatsModels.swift
@@ -61,7 +61,7 @@ struct BookStats: Codable, Equatable {
     let totalChapters: Int
     let completedChapters: Int
     let totalWords: Int
-    let avgChapterWords: Int
+    let avgChapterWords: Double
     let completionRate: Double
     let lastUpdated: String?
 
@@ -82,7 +82,7 @@ struct BookStats: Codable, Equatable {
         self.totalChapters = try c.decodeIfPresent(Int.self, forKey: .totalChapters) ?? 0
         self.completedChapters = try c.decodeIfPresent(Int.self, forKey: .completedChapters) ?? 0
         self.totalWords = try c.decodeIfPresent(Int.self, forKey: .totalWords) ?? 0
-        self.avgChapterWords = try c.decodeIfPresent(Int.self, forKey: .avgChapterWords) ?? 0
+        self.avgChapterWords = try c.decodeIfPresent(Double.self, forKey: .avgChapterWords) ?? 0.0
         self.completionRate = try c.decodeIfPresent(Double.self, forKey: .completionRate) ?? 0.0
         self.lastUpdated = try c.decodeIfPresent(String.self, forKey: .lastUpdated)
     }
